requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pytz>=2024.1
psutil>=5.9.8
requests>=2.31.0
aiohttp>=3.8.5
sqlalchemy>=2.0.0 